Make if thread safe and remove unnecessary funs
    I know, the funs is a pun. But anyway, a lock is added when writing
to a file to prevent two or more lightweight threads (as described by
Google) to corrupt the file.

    Some unnecessary functions were also removed as they where useless
because inside the package it's allowed to acces non exported struct
fields.
